requirements.txt
fusepy
pywebhdfs
